+++ tests/testthat/test-tables_from_seriesid.R
@@ -8,8 +8,10 @@ test_that("correct tableis identified when single series_id supplied", {
     length(single_id$table_no),
     length(single_id$cur_hist)
   )
-  expect_equal(as.character(single_id$table_no[1]),
-               as.character("G1"))
+  expect_equal(
+    as.character(single_id$table_no[1]),
+    as.character("G1")
+  )
 })
 
 test_that("correct table is identified when multiple series_ids from one table are supplied", {
@@ -23,8 +25,10 @@ test_that("correct table is identified when multiple series_ids from one table a
     length(multiple_ids$table_no),
     length(multiple_ids$cur_hist)
   )
-  expect_equal(as.character(multiple_ids$table_no[1]),
-               as.character("G1"))
+  expect_equal(
+    as.character(multiple_ids$table_no[1]),
+    as.character("G1")
+  )
 })
 
 test_that("correct tables are identified when multiple series_ids from different tables are supplied", {
@@ -38,8 +42,10 @@ test_that("correct tables are identified when multiple series_ids from different
     length(multiple_ids$table_no),
     length(multiple_ids$cur_hist)
   )
-  expect_equal(as.character(multiple_ids$table_no),
-               as.character(c("A1", "A1.1", "G1")))
+  expect_equal(
+    as.character(multiple_ids$table_no),
+    as.character(c("A1", "A1.1", "G1"))
+  )
 })
 
 test_that("tables_from_series_id() fails when incorrect series_id(s) supplied", {